requirements.txt
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
notion-client>=2.0.0
slack-sdk>=3.19.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
anthropic>=0.5.0
asyncio>=3.4.3
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0
freezegun>=1.0.0

# Linting and Security
flake8>=4.0.0
black>=22.3.0
bandit>=1.7.0
safety>=2.3.0
isort>=5.10.0
mypy>=1.0.0
pylint>=2.17.0

# Type stubs
types-requests>=2.31.0
types-PyYAML>=6.0.12
types-setuptools>=69.0.0
types-python-dateutil>=2.8.19
types-google-auth>=1.0.0
types-google-auth-oauthlib>=1.0.0
